ai_services/resume-optimizer/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI & NLP
google-generativeai==0.3.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.33.2

# NLP Libraries
nltk==3.8.1
spacy==3.7.2

# PDF Processing
PyPDF2==3.0.1

# Core Libraries
scikit-learn==1.3.2
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.4.2

# HTTP
httpx==0.25.2